Remove RDDIntegrationTest, move all tests involving sRDD into SRDDTest
